refactoring the seperate shacl validates file away
